+++ src/Cwaff-Guns/Gradius.cs
@@ -53,8 +53,8 @@ public class Gradius : CwaffGun
         _RoundLaserProjectile = Items._38Special.DefaultProjectile().Clone(GunData.New(damage: 25f)).ConvertToSpecialtyType<RoundLaser>();
         _WeakseekerProjectile = Items._38Special.CloneProjectile(GunData.New(speed: 300f))
           .SetAllImpactVFX(VFX.CreatePool("lord_projectile_impact_vfx", fps: 60, loops: false, anchor: Anchor.MiddleCenter,
-            emissivePower: 0.5f, emissiveColorPower: 1.5f, emissiveColour: ExtendedColours.vibrantOrange, scale: 0.75f,
-            lightColor: ExtendedColours.vibrantOrange, lightRange: 1.5f, lightStrength: 7.0f))
+            emissivePower: 2f, emissiveColorPower: 10.0f, emissiveColour: new Color(1.0f, 0.85f, 0.5f), emissiveSensitivity: 0.6f,
+            scale: 0.75f, lightColor: ExtendedColours.vibrantOrange, lightRange: 1.5f, lightStrength: 7.0f))
           .AttachTrail("weakseeker_trail", fps: 60, cascadeTimer: C.FRAME, softMaxLength: 0.25f,
             dispersalPrefab: Lazy.DispersalParticles(ExtendedColours.vibrantOrange));