+++ changelog.md
@@ -20,6 +20,7 @@
 - Synchronized position of Stereoscope's VFX with Stereoscope itself while the player is moving
 - Fixed index out of range exception caused by Chroma, Sunderbuss, and Stereoscope failing to look up sprite palette information for modded enemies
 - Improved lighting on trails for projectiles from Aimu Hakurei, Fuel Rod Gun, and Gadulo
+- Improved lighting on impact VFX for Retina, Ki Blast, Wavefront, Heartbreaker, Gradius, Gadulo, Aimu Hakurei, and Jugglernaut
 
 ## 1.31.0 (2026-05-20)